requirements.txt
# LogiFlow ETL Pipeline
# E-Commerce Logistics Data Engineering Platform

# Core dependencies
pandas>=2.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database drivers
psycopg2-binary>=2.9.0  # PostgreSQL

# Data validation
pydantic>=2.0.0

# Logging and monitoring
structlog>=23.0.0

# Orchestration
apache-airflow>=2.7.0

# Streaming (optional, for demo)
# kafka-python>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
